Moved code into main.go
